refactor: provide settings at root module level
