+++ tests/datasets/test_datasets.py
@@ -1,9 +1,9 @@
 import pytest
 
 import rubrix as rb
+from rubrix import TextClassificationSettings, TokenClassificationSettings
 from rubrix.client import api
 from rubrix.client.sdk.commons.errors import AlreadyExistsApiError
-from rubrix.datasets import TextClassificationSettings, TokenClassificationSettings
 
 
 @pytest.mark.parametrize(
